% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_internet}
\alias{check_internet}
\title{Check Internet Connection}
\usage{
check_internet()
}
\value{
Prints a message indicating whether the computer is connected to the internet.
}
\description{
This function checks if the computer is connected to the internet.
If not connected, it stops execution with an error message.
}
