% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_need_libcurl_condathis}
\alias{check_need_libcurl_condathis}
\title{check_libcurl_condathis}
\usage{
check_need_libcurl_condathis()
}
\value{
Boolean
}
\description{
Check if System is Linux and if libcurl version is more than 7.78.0.
}
