% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_status_code}
\alias{check_status_code}
\title{Check the status code of an HTTP response}
\usage{
check_status_code(resp)
}
\arguments{
\item{resp}{An HTTP response object from the httr2 package.}
}
\value{
This function does not return a value. It is used for its side effects.
}
\description{
This function checks the status code of an HTTP response and provides
appropriate messages based on the status.
}
\keyword{internal}
