% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_url}
\alias{check_url}
\title{check_url}
\usage{
check_url(url_to_check)
}
\arguments{
\item{url_to_check}{String of url to check. Note that \verb{http://}  uses port 80
whereas \verb{http://} uses port 443.}
}
\value{
Boolean
}
\description{
Check if url can be read... so if it is up. This function is a present from
Lucio Queiroz.
}
