% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_url_internet}
\alias{check_url_internet}
\title{Check Url and Internet}
\usage{
check_url_internet(url_to_check)
}
\arguments{
\item{url_to_check}{string, url.}
}
\description{
Stop if URL unreachable or not online.
}
