% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr.R
\name{extr_ghs_pubchem}
\alias{extr_ghs_pubchem}
\title{extr_ghs_pubchem}
\usage{
extr_ghs_pubchem(casrn)
}
\arguments{
\item{casrn}{Character vector of carn.}
}
\value{
Dataframe of GHS info.
}
\description{
extr_ghs_pubchem
}
