% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr.R
\name{extr_ice}
\alias{extr_ice}
\title{extr_ice}
\usage{
extr_ice(ids, assays = NULL)
}
\arguments{
\item{ids}{Vector of chemical ids.}

\item{assays}{Vector of assays to extra info of.}
}
\value{
A dataframe.
}
\description{
Extra info from the Integrated Chemical Environment using ICE REST API
}
