% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_tox.R
\name{extr_iris_}
\alias{extr_iris_}
\title{Extract Data from EPA IRIS Database}
\usage{
extr_iris_(
  casrn = NULL,
  cancer_types = c("non_cancer", "cancer"),
  verify_ssl = FALSE,
  ...
)
}
\arguments{
\item{casrn}{A single character string specifying the CASRN for the search.}

\item{cancer_types}{A character vector specifying the types of cancer to include in the search. Must be either "non_cancer" or "cancer".}

\item{verify_ssl}{Boolean to control of SSL should be verified or not.}

\item{...}{Any other arguments to be supplied to \code{req_option} and thus to \code{libcurl}.}
}
\value{
A data frame containing the extracted data.
}
\description{
The \code{extr_iris} function sends a request to the EPA IRIS database to search for information based on a specified keywords and cancer types. It retrieves and parses the HTML content from the response.
Note that if \code{keywords} is not provide all dataset are retrieved.
}
\examples{
\dontrun{
extr_iris("1332-21-4")
}
}
\seealso{
\href{https://cfpub.epa.gov/ncea/iris/search/}{EPA IRIS database}
}
\keyword{internal}
