% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_tox.R
\name{extr_iris_linux_}
\alias{extr_iris_linux_}
\title{extr_iris_linux}
\usage{
extr_iris_linux_(casrn, cancer_types = c("non_cancer", "cancer"))
}
\arguments{
\item{casrn}{A single character string specifying the CASRN for the search.}

\item{cancer_types}{A character vector specifying the types of cancer to include in the search. Must be either "non_cancer" or "cancer".}
}
\description{
extr_iris_linux
}
