% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_pubchem.R
\name{extr_pubchem_fema_}
\alias{extr_pubchem_fema_}
\title{extr_pubchem_fema_}
\usage{
extr_pubchem_fema_(casrn)
}
\arguments{
\item{casrn}{Atomic Vector.}
}
\description{
extr_pubchem_fema_
}
\keyword{internal}
