% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_pubchem.R
\name{extr_pubchem_ghs_}
\alias{extr_pubchem_ghs_}
\title{extr_ghs_pubchem}
\usage{
extr_pubchem_ghs_(casrn)
}
\arguments{
\item{casrn}{Atomic character vector of casrn.}
}
\value{
Dataframe of GHS info.
}
\description{
Extract GHS codes from PubChem. The function relay on the package \code{webchem} to interact with pubchem.
}
