% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr.R
\name{extr_tox}
\alias{extr_tox}
\title{extr_tox}
\usage{
extr_tox(casrn)
}
\arguments{
\item{casrn}{}
}
\value{
List of Dataframes.
}
\description{
Extract toxicological information from PubChem, ICE, Comptox, IRIS.
}
