% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr.R
\name{get_comptox}
\alias{get_comptox}
\title{get_comptox}
\usage{
get_comptox(ids, ...)
}
\arguments{
\item{ids}{A vector of characters where each element is a substance descriptor you wish to query.}

\item{...}{Any other argument of \code{ECOTOXr::websearch_comptox}}
}
\value{
List of data.frames
}
\description{
get_comptox
}
