% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{iris_chem}
\alias{iris_chem}
\title{Iris Chemical Composition}
\format{
An object of class \code{data.frame} with 665 rows and 8 columns.
}
\source{
\href{https://cfpub.epa.gov/ncea/iris/search/index.cfm}{IRIS (Integrated Risk Information System)}
}
\usage{
data(iris_chem)
}
\description{
Dataset containing the chemical composition of iris flowers.
}
\details{
This dataset provides information on the chemical composition of iris flowers.
}
\keyword{datasets}
