% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr.R
\name{items_to_down}
\alias{items_to_down}
\title{items_to_down}
\usage{
items_to_down()
}
\description{
Columns of Main Data of CompTox extracted data. Used in \code{get_comptox}
}
\seealso{
\code{\link[=get_comptox]{get_comptox()}}
}
