% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_ssl}
\alias{set_ssl}
\title{Verify SSL}
\usage{
set_ssl(verify_ssl, ...)
}
\arguments{
\item{verify_ssl}{Boolean.}

\item{...}{Any other arguments to be supplied to \code{req_option}}
}
\value{
Named list.
}
\description{
Verify SSL
}
\keyword{internal}
