% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_internet}
\alias{check_internet}
\title{Check Internet}
\usage{
check_internet(verbose = TRUE)
}
\arguments{
\item{verbose}{Boolean to display messages.}
}
\description{
Wrapper around \code{pingr::is_online} to print message
a better message.
}
\keyword{internal}
