% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr.R
\name{extr_tox}
\alias{extr_tox}
\title{extr_tox}
\usage{
extr_tox(casrn)
}
\arguments{
<<<<<<< HEAD
\item{casrn}{A vector of CASRN.}
=======
\item{casrn}{CASRN in vector.}
>>>>>>> b50926e (fix: checks)
}
\value{
List of dataframes.
}
\description{
Extract toxicological information from PubChem, ICE, Comptox, IRIS.
}
