% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_tox.R
\name{extr_tox}
\alias{extr_tox}
\title{extr_tox}
\usage{
extr_tox(casrn)
}
\arguments{
\item{casrn}{A vector of CASRN.}
}
\value{
List of dataframes.
}
\description{
Extract toxicological information from PubChem, ICE, Comptox, IRIS.
}
