% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{iarc_chem}
\alias{iarc_chem}
\title{IARC Classification}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1115 rows and 7 columns.
}
\source{
\href{https://monographs.iarc.who.int/list-of-classifications/}
}
\usage{
data(iarc_chem)
}
\description{
WHO Agents classified by the IARC Monographs, Volumes 1–135.
}
\keyword{datasets}
