% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{iris_chem}
\alias{iris_chem}
\title{Integrated Risk Information System}
\format{
An object of class \code{data.frame} with 665 rows and 8 columns.
}
\source{
\href{https://cfpub.epa.gov/ncea/iris/search/index.cfm}
}
\usage{
data(iris_chem)
}
\description{
Integrated Risk Information System of US EPA, listing 665 chemicals. Data downloaded in date 05/31/2024
}
\keyword{datasets}
